requirements.txt
requests
pydub
numpy
matplotlib
simpleaudio
pyaudio
gpiozero
smbus2
